requirements.txt
requests
beautifulsoup4
psycopg2-binary
selenium
psycopg2-binary 
python-dotenv
fastapi
pydantic
uvicorn